Split examples into subdirectories
Create the configuration files that are written
in README.md as separate files. These text chunks
are thus duplicated which violates the DRY-principle
(Don't repeat yourself). It would be good to avoid
this duplication in the future, for instance by
generating out HTML documentation in such a way
that the configuration files are included.

Signed-off-by: Erik Sjölund <erik.sjolund@gmail.com>