Add Example 6
This example is similar to Example 5 but here the backend web server is
started with socket activation from a systemd system service with `User=test6`.
No systemd user services are used.
The container running the proxy is currently configured with`--security-opt label=disable`
which means that SELinux is disabled for that container.

Signed-off-by: Erik Sjölund <erik.sjolund@gmail.com>